Add support for store versioning
