+++ test/Microsoft.AspNetCore.Identity.EntityFrameworkCore.Test/UserStoreTest.cs
@@ -14,6 +14,11 @@ using Xunit;
 
 namespace Microsoft.AspNetCore.Identity.EntityFrameworkCore.Test
 {
+    internal class UserStore : UserStore<IdentityUser, IdentityRole, IdentityDbContext, string, IdentityUserClaim<string>, IdentityUserRole<string>, IdentityUserLogin<string>, IdentityUserToken<string>, IdentityRoleClaim<string>>
+    {
+        public UserStore(IdentityDbContext context, IdentityErrorDescriber describer = null) : base(context, describer ?? new IdentityErrorDescriber()) { }
+    }
+
     public class UserStoreTest : IdentitySpecificationTestBase<IdentityUser, IdentityRole>, IClassFixture<ScratchDatabaseFixture>
     {
         private readonly ScratchDatabaseFixture _fixture;
@@ -73,12 +78,12 @@ namespace Microsoft.AspNetCore.Identity.EntityFrameworkCore.Test
 
         protected override void AddUserStore(IServiceCollection services, object context = null)
         {
-            services.AddSingleton<IUserStore<IdentityUser>>(new UserStore<IdentityUser, IdentityRole, IdentityDbContext>((IdentityDbContext)context));
+            services.AddSingleton<IUserStore<IdentityUser>>(new UserStore((IdentityDbContext)context));
         }
 
         protected override void AddRoleStore(IServiceCollection services, object context = null)
         {
-            services.AddSingleton<IRoleStore<IdentityRole>>(new RoleStore<IdentityRole, IdentityDbContext>((IdentityDbContext)context));
+            services.AddSingleton<IRoleStore<IdentityRole>>(new RoleStore<IdentityRole, IdentityDbContext, string, IdentityUserRole<string>, IdentityRoleClaim<string>>((IdentityDbContext)context));
         }
 
         [Fact]
@@ -190,6 +195,20 @@ namespace Microsoft.AspNetCore.Identity.EntityFrameworkCore.Test
             var guid = Guid.NewGuid().ToString();
             var user = new IdentityUser { UserName = "New" + guid };
             IdentityResultAssert.IsSuccess(await manager.CreateAsync(user));
+            IdentityResultAssert.IsSuccess(await manager.DeleteAsync(user));
+        }
+
+        [ConditionalFact]
+        [FrameworkSkipCondition(RuntimeFrameworks.Mono)]
+        [OSSkipCondition(OperatingSystems.Linux)]
+        [OSSkipCondition(OperatingSystems.MacOSX)]
+        public async Task CreateUserSetsCreateDate()
+        {
+            var manager = CreateManager();
+            var guid = Guid.NewGuid().ToString();
+            var user = new IdentityUser { UserName = "New" + guid };
+            IdentityResultAssert.IsSuccess(await manager.CreateAsync(user));
+            Assert.NotNull(await manager.GetCreateDateAsync(user));
             IdentityResultAssert.IsSuccess(await manager.DeleteAsync(user));
         }
